Bug#33384978 Improve cached table stats concurrency
There is a concurrency problem when updating the cached table stats with
changed rows, with several threads updating same table the threads
compete to take the NDB_SHARE::mutex in order to update the cached
"row_count" value.

Rewrite storage of changed rows to use an atomic counter instead of
trying to grab mutex and update the actual shared value. This reduces
the need to serialize the threads. Change to append the number of
changed rows to "row_count" only when taking the stats out of the cache.
Introduce a separate mutex protecting only the cached stats.

Remove usage of zerofilling malloc for creating NDB_SHARE to allow
standard classes to be used as members. This change shows that the
"tuple id range" member traditionally has been initalized to zero rather
than to its invalid value, it is hard to know if that's intentional or
not. Add explicit initialization in order to keep existing zero init
behaviour for tuple id range.

Testing with "sysbench oltp_write_only" using 512 threads and one table
shows 10-20% improvement.

Change-Id: If2f845b6a940e378372a4c91b5e1ed1cc75f9f5e