Bug #33073354 	TOO FEW CHUNK FILES ARE CREATED FOR HYBRID HASH JOIN
Contributed by Oystein Grovlen

For some arbitrary sizes for the join buffer, the number of chunk files
created for hash join is too small.  This means that each file will contain
more rows that can fit in the join buffer.  When this happens the probe
chunks will have to be read multiple times.

The cause for the issue is that by mistake integer division is used when
computing how many chunk files are needed. This is fixed by using floating-
point division.

Change-Id: I1adc4fd35b4bbc578e75f8d8c81be0b2a1dfb7cc