Bug#32906654 ADD NDBINFO TABLE FOR INDEX STATS
Problem:
--------

Since the index statistics system tables are hidden from the MySQL
Server, testing involved a multi-step process in order to check the
status. This comprised of using the ndb_select_all tool to dump the
contents of the ndb_index_stat_head system table to a file and
parsing it. This is error prone and slow from a testing perspective
since the ndb_select_all tool has to establish a connection to the
Cluster on each occasion.

Fix:
----

Implemented a new virtual ndbinfo table: ndbinfo.ndb$index_stats
and a corresponding view: ndbinfo.index_stats. This is done by
retrieving a subset of the information stored in the
ndb_index_stat_head table using NdbApi. The index stats tests
have also been tweaked to use the new view.

Note that the view contains only 3 columns: index_id, index_version,
and sample_version. The ndb_index_stat tool should be used for
a more thorough description.

Change-Id: Iebc79e5705480e27bd660f3c48d770b159946bb7