Bug#33364732: Preliminary fixes for WL#14673 [found_rows, noclose]
Replace the Query_result_delete::found_rows variable with a boolean
variable that tells if any rows have been buffered for delayed
deletion.

The exact count of rows was never used; it was only used for avoiding
unnecessary lookups in the Unique objects if they were known to be
empty. Furthermore, the count was not necessarily correct, since it
would double-count rows for the non-immediate tables if the same row
ID appeared more than once.

Change-Id: Ifb9c433b775ca9f2d937548f7b1b56c234be6661