Bug #32136255 	I_INNODB.TABLE_COMPRESS_* TESTS FAIL ON WINDOWS
This change fixed several bugs with table compression:
1. Bug #32323281 	OS_IS_SPARSE_FILE_SUPPORTED FUNCTION NOT WORKING AS EXPECTED ON WINDOWS IN PB2

ImDisk virtual disk driver we use on Windows does not support newer APIs like Virtual
https://ltr-data.se/opencode.html/#ImDisk section "Compatibility" states:

"ImDisk Virtual Disk Driver is designed to be a small, simple and yet powerful virtual disk driver. It runs on very old versions of Windows NT as well as modern Windows versions. However, because of this compatibility design and because it emulates disk volumes rather than complete disks, [...] you cannot manage things like mount points, drive letters and similar for ImDisk drives using mountvol command line tool or in Disk Management in Windows."

This includes Volume Management API that we use: https://docs.microsoft.com/en-us/windows/win32/FileIO/volume-management-functions

Apparently we don't need to use this API. On all other platforms we just try to punch a hole in a file to check for the sparse file support, and it seems to work just fine on Windows also.

2. `os_file_punch_hole` is not called on Windows. During async file writes `AIOHandler::post_io_processing` is supposed to call `AIOHandler::io_complete` which punches the hole, but it does not, because `/* The punch hole has been done on collect() */`, and Windows don't have `collect`. Fixed Windows async implementation to call `AIOHandler::io_complete`.

3. Holes are not punched for encrypted tablespaces. The `AIOHandler::io_complete` calls `os_file_io_complete` which checks:
```
  } else if (type.punch_hole()) {
    [...]
    /* Nothing to do. */
    if (len == src_len) {
      return (DB_SUCCESS);
```
and indeed the data is compressed from 16K to 8K, but the IO request from double write buffer is done for 8KB only, as it already supplies compressed and encrypted data. So 8KB of 8KB is written and no hole is punched. In normal case it works, because we ask fil subsystem to write out 16KB of data and compress it, so it records 16KB should have been written. Additional parameter for IO methods is added to allow to specify the destination length of the write. Additionally the `file::Block` now keeps track of the encrypted (and possibly compressed) data's size, and there is no need to pass this value through the doublewrite buffer.

4. `table_compress_2` on Windows has a typo that causes the Win32 API usage to be invalid. After fixing it, the test works just fine, re-enabled it for Windows.

Overall, all three `table_compress_*` tests are running correctly and passing on Windows and Linux, removed them from `disabled.def` list.

RB#26370