Bug#33364732: Preliminary fixes for WL#14673 [DeleteRowsIterator, noclose]
Add a new iterator type, DeleteRowsIterator, which corresponds to the
AccessPath type DELETE_ROWS. Most of the code in Query_result_delete
is moved into the iterator.

By having an iterator for the delete operation, we enable EXPLAIN
ANALYZE to see how much time is spent on deletion. Note, however, that
EXPLAIN ANALYZE DELETE still doesn't actually delete the rows. It
probably should; otherwise the timings don't really show how long the
deletion takes. But it's not the right place to enable it in a
refactoring patch. And if we enable it, we should also make EXPLAIN
ANALYZE INSERT and EXPLAIN ANALYZE UPDATE perform inserts and updates
on the target tables.

Change-Id: I5da00b7efad915b68f6f7f90f3693465ba80d5f0