Bug #33409812 Invalid paths found in RPATH for group_replication.so
'check-rpaths' complains about bad RPATH values when building RPMs on
Fedora 35.

The pushbuild "dual" build basically goes like:
 - mkdir debug; cd debug; cmake .. -DWITH_DEBUG=ON; make; cd ..
 - mkdir release; cd release; cmake ..; make; make install

The install from the release directory will pick up debug versions of
mysqld (and rename it to mysqld-debug) and debug versions of all
plugins. These executables are external to the current cmake build,
and so will not be cleaned up with any INSTALL_RPATH property.

Affected executables are mysqld-debug and group_replication.so.  It
also affects fido authentication plugins, iff they are linked with the
bundled libfido2 library.

This patch adds explicit INSTALL(SCRIPT ...) to do the necessary
cleanup of RPATH after intall for mysqld-debug and group_replication.so.
For the fido authentication plugins, we switch to system libfido2 instead.

Change-Id: Ia94c3eb7b29c473ccca0d235568b94d32be7d467