Bug#33364732: Preliminary fixes for WL#14673 [EXPLAIN, noclose]
Fix a regression in how EXPLAIN FORMAT=TREE prints plans for DELETE
statements with impossible WHERE clauses. It was introduced by an
earlier patch for this bug which introduced the DELETE_ROWS access
path type.

Before that patch, such a DELETE statement would display something
like this:

-> Delete from
  -> Zero rows (Impossible WHERE)  (cost=0.00..0.00 rows=0)

After the patch, it no longer showed that it was a DELETE statement,
and simply printed this:

-> Zero rows (Impossible WHERE)  (cost=0.00..0.00 rows=0)

This happened because the code that created the zero rows access path
did not add a DELETE_ROWS access path on top of it. This patch adds
the missing path on top of the zero rows path, and now the output from
EXPLAIN is:

-> Delete from t (buffered)  (cost=0.00..0.00 rows=0)
  -> Zero rows (Impossible WHERE)  (cost=0.00..0.00 rows=0)

(Notice that the description of the top-level path now also includes
the table name and the cost estimate, which were not included even
before the regression.)

Change-Id: I766f09c76bba2797301373ddcf5dd21819be51e6