Bug#33364732: Preliminary fixes for WL#14673 [sql_safe_updates, noclose]
Implement a check for the sql_safe_updates option which looks at the
access paths instead of the JOIN_TABs. This will be needed when adding
support for DELETE statements to the hypergraph optimizer, as the
hypergraph optimizer does not use the JOIN_TAB structure.

The check is moved from Query_result_delete::optimize() to
DeleteRowsIterator::Init() because the access paths have not yet been
created when Query_result_delete::optimize() is called. This also
means that Query_result_delete::optimize() is now empty and can be
removed.

Whereas the original check only raised an error if a full table scan
was used on one of the tables deleted from, the new check also raises
an error if a full index scan is used. This seems to be more in line
with how the option is documented. However, in practice, it does not
make much difference, because the old optimizer is extremely unlikely
to choose a full index scan on one of the tables deleted
from. (Because covering index scans are disabled for tables deleted
from, and also because single-table DELETE uses a light-weight
optimization which does attempt to replace sorts on top of a table
scan with an ordered index scans. Multi-table DELETE does not allow an
ORDER BY clause, so it won't use ordered index scans either.)

This will change when the hypergraph optimizer is used for DELETE
statements. It will run a full optimization of single-table DELETE
statements too, and is more likely to create a plan with an ordered
full index scan in order to avoid a sort. So we add a check for full
index scans now to make it ready to be used by the hypergraph
optimizer.

The patch also adds some extra test coverage for sql_safe_updates mode
with single-table DELETE.

Change-Id: I4ea86254d25df86034c762894cdb3a3d1c691f4e