WL #14807: Index merge in the hypergraph optimizer
Implement index merge (by means of deduplicating row IDs, by sorting
them) in the hypergraph join optimizer. Each OR predicate can give rise
to one or more index merges, but each index merge represents only
a single top-level predicate (this is an unfortunate holdover from
the old join optimizer).

Just like other range index scans, we need to take some care to
understand when we can elide filters and when we cannot; we must keep it
not only if the component scans are inexact, but also in the case where
the different sub-predicates (in ANDs that are not on the top level)
have been split across different candidate index merges.

Change-Id: Icac46af8963a9cb0294c7c21f5b8eebe17f7921c