Bug#33172887 password_last_changed is geting updated on sql node restart
For a user with privilege NDB_STORED_USER, the password_last_changed
timestamp would be updated every time a mysql server restarted.

Fix this in NDB stored grants. When applying a snapshot for a user that
already exists, check the local SHOW CREATE USER and compare it to the
one stored in the snapshot. If the two are already identical, do not
apply the stored CREATE USER statement.

Change-Id: Ieef0faeb2895a57b5b3cb96f8a223c7da33ff311