Bug #33437079 Router tests leave some temporary dirs behind
There are 2 issues in the code:
1) test_logging tests are not using RAII wrapper around the temporary directory
to clean up the created files
2) there are 2 instances of TempDir RAII wrapper, the one in the harness was
not using proper function to delete the directory (remove vs. remove_recursive)

This patch merges the 2 instance of the TempDir wrapper, move is to the harness
and uses in the missing places.

It also uses \tmp directory on unix based OSes for temporary directories
that Router tests create.

RB: 27096