Bug#32213576: MySQL tries to parse string-bound FLOAT as INTEGER and returns wrong error Bug#32787037: Passed string type binding for INTEGER column incorrectly cast to integer.
This is a regression from WL#9384. In this worklog, dynamic parameter
handling was refactored so that the derived data type for parameters was
determined based on context, e.g in a comparison <int_col> = ?,
the parameter was given the same type as the (integer) column it was
compared to. To preserve compatibility with existing MySQL applications,
if a decimal or float value was supplied as parameter, the statement
would be automatically reprepared with new type assigned to the
parameter, based on the actual value. This preserved compatibility
for numeric parameters. However, if a string parameter was supplied,
it would still be interpreted as an integer (the resolved data type).
This is not compatible with older MySQL versions that would detect
the actual type of the value. This means that if <int_col> = ? is
executed with parameter value '1.7', only the integer part of the string
would be used, thus effectively the comparison would be <int_col> = 1.

The solution builds on the refactoring done in bug#32915973.
When a string parameter is supplied, the parameter is analyzed to
determine if it is an integer, a decimal or a float value, and the
"actual" data type of the parameter is updated accordingly.
The actual type is later compared to the resolved type and if
incompatible, the statement is reprepared with the new actual type.
Thus, the statement above is evaluated as int_col> = 1.7, and the
comparison evaluated using decimal numbers.

The fix implies some test case changes, in particular in
main.func_comparison, where some comparisons that were previously
evaluated using float numbers, and thus could be inexact for large
integers, are now evaluated as integer operations with better accuracy.

Change-Id: Iaf851a0456eadab9e7b0c1a1e8061d1e57e957c7