Bug#33414235: Item_row::illegal_method_call const: Assertion `0' failed.
If the first argument given to JSON_TABLE is a row instead of a single
JSON value, an assertion is raised when trying to evaluate the row
expression.

Fixed by raising an error during resolving if the first argument is a
row, so that the row expression is never evaluated.

Change-Id: I469ef87029cabdbff3d3b4b85d8c36053729190a