Bug#33394135 - adding CHECK constraint with alter table CHANGE                column name is not working
Adding check constraint to an existing column with ALTER TABLE
CHANGE column and MODIFY column is not working.

Check constraint information obtained from the ALTER TABLE CHANGE
column and MODIFY column is ignored by mistake. Because of which
check constraint is not created. Adjusted code use check constraint
information obtained from the CHANGE and MODIFY column to fix
this issue.

Change-Id: I4c953698a20e5c61caa9e2a3c251090104f5b8f7