Bug#33428567 Fix server message when usign relay log recovery and GTID_ONLY = 1
When GTID_ONLY is enabled for a channel and a restart occurs where
relay log recovery is used, a message was printed about the adjustment
of positions based on outdated data.

Since positions are not needed for relay log recovery when GTIDs are
used in the replica, a new message with no position information is
printed when GTID_ONLY = 1.

ReviewBoard: 27110