Bug #30617496 BAD ITEM REF FROM CORRELATED SUBQUERY TO OUTER DISTINCT TABLE
Contributed by Song Zhibai

A DML query contains an outer DISTINCT query which has a subquery
inside a HAVING clause. The inner subquery attempts to use a column
reference for a column from the outer DISTINCT query. This should be
allowed only if the subquery is used somewhere out of HAVING or the
outer SELECT does not use grouping. However, the query is allowed to
run even though neither of these conditions are met.

The column reference check is expanded to detect the invalid column
reference and return an error.

Change-Id: I39e9feca3581132f9f0600ba4a1f52fafef1b2e0