Bug #33491886: Nested loop joins allow very expensive inner sides in the hypergraph optimizer
If the outer side of a join is (mis)estimated to have very low
cardinalities, such as 1e-5, the hypergraph optimizer thinks it can put
almost arbitrarily expensive things on the inner side, such as a full
table scan. Obviously, the former shouldn't happen, but until we have
great row estimation, add a workaround where we do ceil() on the
cardinality.

Change-Id: I00041138937827e7f190c92b5e018c77feaf05db