Bug#33452674 join_pushdown_*.test: Use regex filtering to shorten IN-value lists in *.result
Several test cases in the join_pushdown_*.test generated a very long IN-value list.
That cluttered up the *.result files for the test cases to the point that they
became unreadable.

Patch use regex filtering to shorten the IN-lists being reported to
the *.result files such that only the first and last item in the
IN-value list is reported, seperated by '....'.'

Reviewed By: Arnab Ray <arnab.r.ray@oracle.com>

Change-Id: I4764f59679b5118c10dd55ce1b35ecd44a70af1c