Bug#33066458: TEST_IF_SKIP_SORT_ORDER: ASSERTION `SELECT_LIMIT > 0' FAILED.
In a calculation of the maximum number of rows that are likely to be
read from a scan, the intermediate result was a double that could
become greater than the maximum allowed value for a 64-bit unsigned
integer. This triggered undefined behaviour when converting the
intermediate double value to an integer, which in some cases could
lead to assert failures.

Fixed by clamping the result in the range [1, UINT64_MAX]. (The lower
bound was already enforced. The patch adds the upper bound.)

Change-Id: I083b76341663295647629206fe8b462d0524a604