Bug#32858783 ASSERTION `NULLPTR != DYNAMIC_CAST<TARGET>(ARG)' FAILED.
This error happens for statements on the following form:

INSERT INTO <view> VALUE(<row>) AS <row alias>(<identifier list>)

When executing such statements, the server calls
Sql_cmd_insert_base::prepare_values_table() in order to prepare the
derived table created as a VALUES alias. This function populates
Sql_cmd_insert_base.values_field_list with Item_field objects pointing
to the fields the underlying table. When we insert into a view rather
than a table, there is a missing real_item() transform. This is
needed to map from Item_view_ref (pointing to the view column) to
Item_field (the corresponding field in the underlying table).

This commit adds the missing real_item() call, and a test case.

Change-Id: I8e711f7c61e3e3c9539fd311c42283d7bb8cd6bd