Bug#33066455: REGRESSION: ACCESSPATH* MOVECOMPOSITEITERATORSFROMTABLEPATH(ACCESSPATH*): ASSERTION `FALSE' FAILED.
Queries that use UNION and LIMIT 0 may trigger an assert failure in
debug builds.

The problems started with the fix for bug#32788576, which tries to
move composite paths out of MATERIALIZE paths. The fix did not account
for the possibility of having a MATERIALIZE path on top of a ZERO_ROWS
path, and an assertion is raised if we have such a construction.

It is desirable to move ZERO_ROWS on top if MATERIALIZE too, so that
we don't create a temporary table if there is nothing to materialize.
This would usually be done by making the MATERIALIZE path a child of
the ZERO_ROWS path, and replacing the table_path inside the
MATERIALIZE path with the original child of the ZERO_ROWS path. In
many cases, and in the failing one in particular, ZERO_ROWS paths
don't have a child, so we don't have anything to put into table_path.
We solve this by simply dropping the MATERIALIZE path and replacing it
with the ZERO_ROWS path. We can do this because the ZERO_ROWS path
does not execute any of its child paths anyway, so anything below a
ZERO_ROWS path can safely be removed from the query plan.

Change-Id: I51c5847d67db2799dee06368139ed0fbf2655da9