Bug#33321787: Possible SIGSEGV in set_parse_error_message
Problem: Code inspection showed used of strncpy without ensuring that
the last byte of the buffer being copied into is a null-byte. This looks
bad, but in this case was hardly a real problem, as the argument passed
in was always either taken from a buffer of the same size
as we are copying into, or a string literal less than the size of the
buffer.

Solution: Use snprintf, as is done in Trigger::set_parse_error_message(),
which makes sure the result is valid even if truncated.

Change-Id: I866b7dbbfa53316f5ca3aa63af94ca5cebde9909