Bug#33311645 Fuzzy costs in the hypergraph join optimizer
The hypergraph optimizer sometimes keeps many access paths around that
are very similar, and most likely redundant. In order to reduce the
number of access paths we introduce fuzziness into the comparison that
the optimizer uses to determine which access paths to keep.

When performing a fuzzy comparison of two access paths on a numerical
property such as cost, we consider the paths identical unless their
relative difference in cost exceeds a threshold that we set to 1%.

A proposed path is rejected by the optimizer if it compares identical or
worse to an existing path across all properties (cost, ordering, etc.).
By adding fuzziness to the comparison of the numerical properties of
access paths (cost, initialization cost, rescan cost, and output rows)
we reject proposed paths that do not represent a significant improvement
or difference to existing paths, thus reducing the number of paths kept
by the optimizer.

One downside of using fuzzy comparisons is that a proposed access path
can be rejected if it is no worse than an existing path across all
properties, and slightly better in at least one numerical property.
We implement additional checks to ensure that such paths are kept.

Change-Id: Ia27f5f0708d6ebc2a9a403b0f72d10d918019b4b