Bug#33242407 - INNODB: SIG11 IN calc_row_difference at ha_innodb.cc on select
In the case of a query that uses temporary table for aggregation,
if the group-by field value is already present, the row is updated.
Otherwise, a new row is inserted into the temporary table.

The temporary table may exhaust the memory for an update
operation. In such a case the table is moved to disk and the
update operation is attempted again.

When re-trying the update on the newly created table, the data
prepared before the table was moved to disk is used. In case of
blob columns, the update record holds a pointer to the blob which
was fetched when the temporary table was not moved to disk.

This use of the stale blob pointer may cause the server to exit.

Fix:
The update record is initialized again after the table is moved
to disk and the functions are re-evaluated.

Change-Id: I22b6d9de8a6c708e9ac68b1bd1a67c3902ea3e68