Bug#33157531 ZDI-CAN-14487: Oracle MySQL Cluster Data Node Improper Validation of Array Index ....
Several of the signals received by the SPJ block contain an 'i-pointer', which is
used to look up real memory object to be operated upon.

The ::getPtr() method used to translate between such 'i-pointers' and real pointers,
offers very limited validations of the 'i-pointers', thus we can end up in different
'out of bounds' situations if the received pointers are garbage.

This patch introduce an additional hash-structure for the TreeNode and ScanFraghandle
objects which can be externally refered from such i-pointers. Whenever translating
an incomming i-pointer to a real pointer we noe use these hash structures to find
the real object, instead of using the somewhat unsafe getPtr(). That ensures that
only signals containing an i-pointer known to SPJ will be able to locate the memory
object.

If the received i-pointer can not found via the hash lists a 'ndbrequire' will fail
and the datanode crash (Fail fast, as we usually do in such situations)

Note that we only apply this mechanism for such pointers received in signals.
There are also several SPJ internal usage of these 'i-pointers', but they are
assumed to be trusted.

Reviewed By: Mauritz Sundell <mauritz.sundell@oracle.com>

Change-Id: I9d2b925b712ea328fb3ea1db84aa24e9c7fbb7de