Bug #33406701 	[MY-011879] [InnoDB] Failed to set NUMA memory policy of buffer pool page frames to MPOL_INTERLEAVE (error: Invalid argument)
This is a regression caused by `Bug #32714144  UT0NEW LARGE-PAGE SUPPORT SHOULD ENCAPSULATE THE PFS METADATA`
The buffer pool is allocated in chunks. If NUMA is enabled the memory allocated to chunks is binded to NUMA node with the help of `mbind()` call. The call will only succeed if the memory allocated is aligned to the page size.
In this particular scenario we are using the large pages provided by the kernel and its page size is greater than default system page size. On x64 it can be 2 MB, and the memory allocated to chunks should be aligned to 2MB.
The change of `Bug #32714144` tried to use the first CPU_PAGE_SIZE (in our case 4096 bytes) bytes of the chunk memory for storing PFS metadata and passed `chunk->mem + CPU_PAGE_SIZE` to the `mbind()` call.
When using normal page size `mbind()`will not cause a problem because the address is still aligned to the system page size but when using the large page size it fails and causes NUMA to not be in interleaved mode.
Similar problem occurs with `madvice()` call, for the same reasons. The `madvice()` can be activated by usage of `--innodb_buffer_pool_in_core_file=OFF --core-file`

This change fixes it by providing a low level function call for `mbind()` and `madvice()` to retrieve the original memory chunk pointer allocated by OS.
Also, some assertions are added to make sure the correct flavour of methods are used that suit the `malloc_*page` call used to obtain the pointer. These found a small bug in `buf_chunk_init()`.

RB#27144