BUG#32803050 REPLICA WITH FEDERATED TABLE CRASHES IN HA_FEDERATED::RND_POS WHEN APPLYING ROW EVENTS
Description: When executing Update/Delete with indexes on a table with
Row based replication setup and the federated table on slave, The Slave
crashes.

Analysis: Slave crashes when trying to get the reference to the row
to be update or deleted on federated table using position() method.
The crash occurs due to assert(stored_result) in position method, since
the stored_result is still empty. Server should do index_scan and
update the stored_result and then call for position() and ha_rnd_pos()
methods to get the ref to the row stored. But in this scenario the
server calls for position() method without scanning the table resulting
in crash at assert.

Fix: The fix is added to fetch the position or reference of the row
only when stored_result is not empty.

RB: 26755