Bug#33157531 'Improper Validation of Array Index'
Follow up patch to fix Windows build break

Windows build failed due to hitting  ASSERT_TYPE_HAS_CONSTRUCTOR(T) in DLMHashTable c'tor.
That assert is only defined on Windows platforms as (!__has_trivial_constructor(T)).
It also turns out that this is the only usage of that ASSERT.

Inspecting class DLMHashTable we do not find that this assert provide any securit at all:
The member variables in the 'T' class used by DLMHashTable methods are always set
by the methods themself before relying on their contents.

This patch remove the usage of this ASSERT as well as the definition of the (now unused)
ASSERT itself.

Reviewed By: Mauritz Sundell <mauritz.sundell@oracle.com>

Change-Id: Ib3718efb14a2c57554ecb0c023d50b932aa4373d